requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
cohere==5.13.3
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.86.0
crewai-tools==0.17.0
cryptography==43.0.3
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
durationpy==0.9
embedchain==0.1.125
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.6
fastavro==1.9.7
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-aiplatform==1.74.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
instructor==1.7.0
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.6.1
json_repair==0.30.3
jsonpatch==1.33
jsonpickle==4.0.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
lancedb==0.17.0
langchain==0.3.11
langchain-cohere==0.3.3
langchain-community==0.3.11
langchain-core==0.3.24
langchain-experimental==0.3.3
langchain-ollama==0.2.1
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langsmith==0.1.147
litellm==1.55.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.34
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.4
onnxruntime==1.20.1
openai==1.57.3
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.4
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pylance==0.20.0
pypdf==5.1.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.390
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
pywin32==308
PyYAML==6.0.2
qdrant-client==1.12.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
schema==0.7.7
selenium==4.27.1
setuptools==75.6.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.1.0
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uv==0.5.8
uvicorn==0.32.1
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
